requirements.txt
# Core ML and NLP libraries
transformers>=4.28.0
torch>=1.13.0
tensorflow>=2.10.0
numpy>=1.20.0
pandas>=1.4.0
scikit-learn>=1.0.0
nltk>=3.6.0
spacy>=3.5.0
gensim>=4.2.0

# Visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.4.0

# Model persistence and utilities
joblib>=1.1.0
tqdm>=4.62.0
shap>=0.40.0
lime>=0.2.0

# Image and audio processing
pillow>=9.0.0
opencv-python-headless>=4.5.0
librosa>=0.9.0
soundfile>=0.10.0

# Advanced ML models
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0

# Web framework and API
streamlit>=1.20.0
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.9.0

# Database and storage
sqlalchemy>=1.4.0
redis>=4.0.0
pymongo>=4.0.0

# Monitoring and logging
prometheus-client>=0.14.0
structlog>=22.0.0
sentry-sdk>=1.9.0

# Security and validation
cryptography>=3.4.0
python-jose>=3.3.0
passlib>=1.7.0

# Async and performance
asyncio
aiofiles>=0.8.0
httpx>=0.23.0

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.19.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Utilities
requests>=2.28.0
python-dotenv>=0.19.0
click>=8.0.0
rich>=12.0.0
